+++ packages/expo-secure-store/android/src/main/java/expo/modules/securestore/AuthenticationPrompt.kt
@@ -13,13 +13,13 @@ import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
 import kotlin.coroutines.suspendCoroutine
 
-class AuthenticationPrompt(private val currentActivity: FragmentActivity, context: Context, title: String, isUserPresenceRequired: Boolean) {
-  private var authType: Int = if (isUserPresenceRequired) BIOMETRIC_STRONG or DEVICE_CREDENTIAL else BIOMETRIC_STRONG
+class AuthenticationPrompt(private val currentActivity: FragmentActivity, context: Context, title: String, isDeviceCredentialsRequired: Boolean) {
+  private var authType: Int = if (isDeviceCredentialsRequired) BIOMETRIC_STRONG or DEVICE_CREDENTIAL else BIOMETRIC_STRONG
   private var executor: Executor = ContextCompat.getMainExecutor(context)
   private var promptInfo = PromptInfo.Builder().apply {
     setTitle(title)
     setAllowedAuthenticators(authType)
-    if (!isUserPresenceRequired) {
+    if (!isDeviceCredentialsRequired) {
       setNegativeButtonText(context.getString(android.R.string.cancel))
     }
   }.build()